@@ -1,18 +1,18 @@
 name: upgrader
 description: Flutter package for prompting users to upgrade when there is a newer version of the app in the store.
-version: 9.0.0
+version: 10.3.0
 homepage: https://github.com/larryaasen/upgrader
 
 environment:
   sdk: '>=3.1.0 <4.0.0'
-  flutter: ">=3.13.1"
+  flutter: ">=3.16.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # From fluttercommunity.dev: Get current device information from within the Flutter application.
-  device_info_plus: '>=8.1.0 <10.0.0'
+  device_info_plus: '>=8.1.0 <11.0.0'
 
   # Pure Dart library for HTML5 parsing
   html: ^0.15.3
@@ -30,7 +30,7 @@ dependencies:
   os_detect: ^2.0.1
 
   # From Flutter Community: Provides an API for querying information about an application package.
-  package_info_plus: '>=4.0.1 <6.0.0'
+  package_info_plus: '>=4.0.1 <9.0.0'
 
   # From Flutter Team: Wraps NSUserDefaults (on iOS) and SharedPreferences (on Android).
   shared_preferences: '>=2.1.1 <3.0.0'
