@@ -1,8 +1,9 @@
 /*
- * Copyright (c) 2019-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2019-2024 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
 import 'package:upgrader/upgrader.dart';
 
 void main() async {
@@ -11,23 +12,21 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  // On Android, setup the Appcast.
-  // On iOS, the default behavior will be to use the App Store version of
-  // the app, so update the Bundle Identifier in example/ios/Runner with a
-  // valid identifier already in the App Store.
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  MyApp({Key? key}) : super(key: key);
+  MyApp({super.key});
 
   static const appcastURL =
       'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
   final upgrader = Upgrader(
-    appcastConfig:
-        AppcastConfiguration(url: appcastURL, supportedOS: ['android']),
-    debugLogging: true,
-    minAppVersion: '1.1.0',
+    client: http.Client(),
+    clientHeaders: {'header1': 'value1'},
+    storeController: UpgraderStoreController(
+      onAndroid: () => UpgraderAppcastStore(appcastURL: appcastURL),
+      oniOS: () => UpgraderAppcastStore(appcastURL: appcastURL),
+    ),
   );
 
   @override
@@ -35,10 +34,10 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Upgrader Example',
       home: Scaffold(
-          appBar: AppBar(title: Text('Upgrader Example')),
+          appBar: AppBar(title: const Text('Upgrader Appcast Example')),
           body: UpgradeAlert(
             upgrader: upgrader,
-            child: Center(child: Text('Checking...')),
+            child: const Center(child: Text('Checking...')),
           )),
     );
   }
