@@ -11,25 +11,21 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  // On Android, setup the Appcast.
-  // On iOS, the default behavior will be to use the App Store version of
-  // the app, so update the Bundle Identifier in example/ios/Runner with a
-  // valid identifier already in the App Store.
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Upgrader Example',
       home: Scaffold(
-        appBar: AppBar(title: Text('Upgrader Cupertino Example')),
+        appBar: AppBar(title: const Text('Upgrader Cupertino Example')),
         body: UpgradeAlert(
           dialogStyle: UpgradeDialogStyle.cupertino,
-          child: Center(child: Text('Checking...')),
+          child: const Center(child: Text('Checking...')),
         ),
       ),
     );
