@@ -11,31 +11,30 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  // On Android, setup the Appcast.
-  // On iOS, the default behavior will be to use the App Store version of
-  // the app, so update the Bundle Identifier in example/ios/Runner with a
-  // valid identifier already in the App Store.
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  MyApp({Key? key}) : super(key: key);
+  MyApp({super.key});
 
   static const appcastURL =
       'https://raw.githubusercontent.com/larryaasen/upgrader/master/test/testappcast.xml';
   final upgrader = Upgrader(
-      appcastConfig:
-          AppcastConfiguration(url: appcastURL, supportedOS: ['android']));
+    storeController: UpgraderStoreController(
+        onAndroid: () => UpgraderAppcastStore(appcastURL: appcastURL)),
+    debugLogging: true,
+    minAppVersion: '1.1.0',
+  );
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Upgrader Example',
       home: Scaffold(
-          appBar: AppBar(title: Text('Upgrader Appcast Example')),
+          appBar: AppBar(title: const Text('Upgrader Example')),
           body: UpgradeAlert(
             upgrader: upgrader,
-            child: Center(child: Text('Checking...')),
+            child: const Center(child: Text('Checking...')),
           )),
     );
   }
