@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2019-2024 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/material.dart';
@@ -11,28 +11,23 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  // On Android, the default behavior will be to use the Google Play Store
-  // version of the app.
-  // On iOS, the default behavior will be to use the App Store version of
-  // the app, so update the Bundle Identifier in example/ios/Runner with a
-  // valid identifier already in the App Store.
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  MyApp({Key? key}) : super(key: key);
+  MyApp({super.key});
 
-  final upgrader = MyUpgrader();
+  final _upgrader = MyUpgrader();
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Upgrader Subclass Example',
       home: Scaffold(
-          appBar: AppBar(title: Text('Upgrader Subclass Example')),
+          appBar: AppBar(title: const Text('Upgrader Subclass Example')),
           body: UpgradeAlert(
-            upgrader: upgrader,
-            child: Center(child: Text('Checking...')),
+            upgrader: _upgrader,
+            child: const Center(child: Text('Checking...')),
           )),
     );
   }
