@@ -1,11 +1,14 @@
 /*
- * Copyright (c) 2021-2023 Larry Aasen. All rights reserved.
+ * Copyright (c) 2021-2024 Larry Aasen. All rights reserved.
  */
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:upgrader/upgrader.dart';
 
+import 'upgrade_messages.dart';
+import 'upgrade_state.dart';
+import 'upgrader.dart';
 
 /// There are two different dialog styles: Cupertino and Material
 enum UpgradeDialogStyle { cupertino, material }
@@ -18,8 +21,7 @@ class UpgradeAlert extends StatefulWidget {
   UpgradeAlert({
     super.key,
     Upgrader? upgrader,
-    this.isDark = false,
-    this.canDismissDialog = false,
+    this.barrierDismissible = false,
     this.dialogStyle = UpgradeDialogStyle.material,
     this.onIgnore,
     this.onLater,
@@ -32,13 +34,15 @@ class UpgradeAlert extends StatefulWidget {
     this.dialogKey,
     this.navigatorKey,
     this.child,
+    this.isDark = false,
   }) : upgrader = upgrader ?? Upgrader.sharedInstance;
 
   /// The upgraders used to configure the upgrade dialog.
   final Upgrader upgrader;
 
-  /// Can alert dialog be dismissed on tap outside of the alert dialog. Not used by [UpgradeCard]. (default: false)
-  final bool canDismissDialog;
+  /// The `barrierDismissible` argument is used to indicate whether tapping on the
+  /// barrier will dismiss the dialog. (default: false)
+  final bool barrierDismissible;
 
   /// The upgrade dialog style. Used only on UpgradeAlert. (default: material)
   final UpgradeDialogStyle dialogStyle;
@@ -54,9 +58,7 @@ class UpgradeAlert extends StatefulWidget {
   /// Return false when the default behavior should not execute.
   final BoolCallback? onUpdate;
 
-  /// Called when the user taps outside of the dialog and [canDismissDialog]
-  /// is false. Also called when the back button is pressed. Return true for
-  /// the screen to be popped.
+  /// Called to determine if the dialog blocks the current route from being popped.
   final BoolCallback? shouldPopScope;
 
   /// Hide or show Ignore button on dialog (default: true)
@@ -101,29 +103,30 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   /// Describes the part of the user interface represented by this widget.
   @override
   Widget build(BuildContext context) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: build UpgradeAlert');
     }
 
     return StreamBuilder(
-      initialData: widget.upgrader.evaluationReady,
-      stream: widget.upgrader.evaluationStream,
-      builder:
-          (BuildContext context, AsyncSnapshot<UpgraderEvaluateNeed> snapshot) {
+      initialData: widget.upgrader.state,
+      stream: widget.upgrader.stateStream,
+      builder: (BuildContext context, AsyncSnapshot<UpgraderState> snapshot) {
         if ((snapshot.connectionState == ConnectionState.waiting ||
                 snapshot.connectionState == ConnectionState.active) &&
-            snapshot.data != null &&
-            snapshot.data!) {
-          if (widget.upgrader.debugLogging) {
-            print("upgrader: need to evaluate version");
-          }
+            snapshot.data != null) {
+          final upgraderState = snapshot.data!;
+          if (upgraderState.versionInfo != null) {
+            if (widget.upgrader.state.debugLogging) {
+              print("upgrader: need to evaluate version");
+            }
 
-          if (!displayed) {
-            final checkContext = widget.navigatorKey != null &&
-                    widget.navigatorKey!.currentContext != null
-                ? widget.navigatorKey!.currentContext!
-                : context;
-            checkVersion(context: checkContext);
+            if (!displayed) {
+              final checkContext = widget.navigatorKey != null &&
+                      widget.navigatorKey!.currentContext != null
+                  ? widget.navigatorKey!.currentContext!
+                  : context;
+              checkVersion(context: checkContext);
+            }
           }
         }
         return widget.child ?? const SizedBox.shrink();
@@ -132,11 +135,10 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   }
 
   /// Will show the alert dialog when it should be dispalyed.
-  /// Only called by [UpgradeAlert] and not used by [UpgradeCard].
   void checkVersion({required BuildContext context}) {
     final shouldDisplay = widget.upgrader.shouldDisplayUpgrade();
-    if (widget.upgrader.debugLogging) {
-      print('upgrader: shouldDisplayReleaseNotes: shouldDisplayReleaseNotes');
+    if (widget.upgrader.state.debugLogging) {
+      print('upgrader: shouldDisplayReleaseNotes: $shouldDisplayReleaseNotes');
     }
     if (shouldDisplay) {
       displayed = true;
@@ -150,7 +152,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
           message: widget.upgrader.body(appMessages),
           releaseNotes:
               shouldDisplayReleaseNotes ? widget.upgrader.releaseNotes : null,
-          canDismissDialog: widget.canDismissDialog,
+          barrierDismissible: widget.barrierDismissible,
           messages: appMessages,
         );
       });
@@ -158,7 +160,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   }
 
   void onUserIgnored(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: ignore');
     }
 
@@ -175,7 +177,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   }
 
   void onUserLater(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: later');
     }
 
@@ -188,7 +190,7 @@ class UpgradeAlertState extends State<UpgradeAlert> {
   }
 
   void onUserUpdated(BuildContext context, bool shouldPop) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: button tapped: update now');
     }
 
@@ -220,10 +222,10 @@ class UpgradeAlertState extends State<UpgradeAlert> {
     required String? title,
     required String message,
     required String? releaseNotes,
-    required bool canDismissDialog,
+    required bool barrierDismissible,
     required UpgraderMessages messages,
   }) {
-    if (widget.upgrader.debugLogging) {
+    if (widget.upgrader.state.debugLogging) {
       print('upgrader: showTheDialog title: $title');
       print('upgrader: showTheDialog message: $message');
       print('upgrader: showTheDialog releaseNotes: $releaseNotes');
@@ -233,46 +235,39 @@ class UpgradeAlertState extends State<UpgradeAlert> {
     widget.upgrader.saveLastAlerted();
 
     showDialog(
-      barrierDismissible: canDismissDialog,
+      barrierDismissible: barrierDismissible,
       context: context,
       builder: (BuildContext context) {
-        return WillPopScope(
-          onWillPop: () async => onWillPop(),
-          child: AppUpdateDialog(
-            isDark: widget.isDark,
-            messages: messages,
-            releaseNotes: releaseNotes,
-            ignoreCallback:
-                widget.showIgnore ? () => onUserIgnored(context, true) : null,
-            laterCallback:
-                widget.showLater ? () => onUserLater(context, true) : null,
-            updateCallback: () =>
-                onUserUpdated(context, !widget.upgrader.blocked()),
+        return PopScope(
+          canPop: onCanPop(),
+          onPopInvoked: (didPop) {
+            if (widget.upgrader.state.debugLogging) {
+              print('upgrader: showTheDialog onPopInvoked: $didPop');
+            }
+          },
+          child: alertDialog(
+            key,
+            title ?? '',
+            message,
+            releaseNotes,
+            context,
+            widget.dialogStyle == UpgradeDialogStyle.cupertino,
+            messages,
           ),
         );
-        // alertDialog(
-        //   key,
-        //   title ?? '',
-        //   message,
-        //   releaseNotes,
-        //   context,
-        //   widget.dialogStyle == UpgradeDialogStyle.cupertino,
-        //   messages,
-        // ));
       },
     );
   }
 
-  /// Called when the user taps outside of the dialog and [canDismissDialog]
-  /// is false. Also called when the back button is pressed. Return true for
-  /// the screen to be popped. Defaults to false.
-  bool onWillPop() {
-    if (widget.upgrader.debugLogging) {
-      print('upgrader: onWillPop called');
+  /// Determines if the dialog blocks the current route from being popped.
+  /// Will return the result from [shouldPopScope] if it is not null, otherwise it will return false.
+  bool onCanPop() {
+    if (widget.upgrader.state.debugLogging) {
+      print('upgrader: onCanPop called');
     }
     if (widget.shouldPopScope != null) {
       final should = widget.shouldPopScope!();
-      if (widget.upgrader.debugLogging) {
+      if (widget.upgrader.state.debugLogging) {
         print('upgrader: shouldPopScope=$should');
       }
       return should;
