@@ -1,6 +1,4 @@
-/*
- * Copyright (c) 2019-2024 Larry Aasen. All rights reserved.
- */
+// Copyright (c) 2024 Larry Aasen. All rights reserved.
 
 import 'package:flutter/material.dart';
 import 'package:upgrader/upgrader.dart';
@@ -11,24 +9,23 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  // On Android, the default behavior will be to use the Google Play Store
-  // version of the app.
-  // On iOS, the default behavior will be to use the App Store version of
-  // the app, so update the Bundle Identifier in example/ios/Runner with a
-  // valid identifier already in the App Store.
-  runApp(const MyApp());
+  runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({super.key});
+  MyApp({super.key});
+
+  final _upgrader = Upgrader(
+      debugLogging: true, durationUntilAlertAgain: const Duration(seconds: 10));
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Upgrader Example',
+      title: 'Upgrader Example - Alert Again',
       home: UpgradeAlert(
+        upgrader: _upgrader,
         child: Scaffold(
-          appBar: AppBar(title: const Text('Upgrader Example')),
+          appBar: AppBar(title: const Text('Upgrader Example - Alert Again')),
           body: const Center(child: Text('Checking...')),
         ),
       ),
