@@ -9,27 +9,28 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  MyApp({super.key});
+  const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Upgrader Example',
       home: MyUpgradeAlert(
-          child: Scaffold(
-        appBar: AppBar(title: Text('Upgrader Alert Theme Example')),
-        body: Center(child: Text('Checking...')),
-      )),
+        child: Scaffold(
+          appBar: AppBar(title: const Text('Upgrader Alert Theme Example')),
+          body: const Center(child: Text('Checking...')),
+        ),
+      ),
     );
   }
 }
 
 class MyUpgradeAlert extends UpgradeAlert {
-  MyUpgradeAlert({super.upgrader, super.child});
+  MyUpgradeAlert({super.key, super.upgrader, super.child});
 
   /// Override the [createState] method to provide a custom class
   /// with overridden methods.
@@ -49,8 +50,16 @@ class MyUpgradeAlertState extends UpgradeAlertState {
       UpgraderMessages messages) {
     return Theme(
       data: ThemeData(
-        dialogTheme: DialogTheme(
-            titleTextStyle: TextStyle(color: Colors.red, fontSize: 48.0)),
+        dialogTheme: const DialogTheme(
+          titleTextStyle: TextStyle(color: Colors.red, fontSize: 48),
+          contentTextStyle: TextStyle(color: Colors.green, fontSize: 18),
+        ),
+        textButtonTheme: const TextButtonThemeData(
+          style: ButtonStyle(
+            // Change the color of the text buttons.
+            foregroundColor: MaterialStatePropertyAll(Colors.orange),
+          ),
+        ),
       ),
       child: super.alertDialog(
           key, title, message, releaseNotes, context, cupertino, messages),
