@@ -24,8 +24,8 @@ class MyApp extends StatelessWidget {
       home: MyUpgradeAlert(
           upgrader: upgrader,
           child: Scaffold(
-            appBar: AppBar(title: Text('Upgrader Custom Alert Example')),
-            body: Center(child: Text('Checking...')),
+            appBar: AppBar(title: const Text('Upgrader Custom Alert Example')),
+            body: const Center(child: Text('Checking...')),
           )),
     );
   }
@@ -35,22 +35,17 @@ class MyUpgrader extends Upgrader {
   MyUpgrader({super.debugLogging});
 
   @override
-  bool isTooSoon() {
-    return super.isTooSoon();
-  }
-
-  @override
   bool isUpdateAvailable() {
-    final appStoreVersion = currentAppStoreVersion;
+    final storeVersion = currentAppStoreVersion;
     final installedVersion = currentInstalledVersion;
-    print('appStoreVersion=$appStoreVersion');
+    print('storeVersion=$storeVersion');
     print('installedVersion=$installedVersion');
     return super.isUpdateAvailable();
   }
 }
 
 class MyUpgradeAlert extends UpgradeAlert {
-  MyUpgradeAlert({super.upgrader, super.child});
+  MyUpgradeAlert({super.key, super.upgrader, super.child});
 
   /// Override the [createState] method to provide a custom class
   /// with overridden methods.
@@ -66,7 +61,7 @@ class MyUpgradeAlertState extends UpgradeAlertState {
     required String? title,
     required String message,
     required String? releaseNotes,
-    required bool canDismissDialog,
+    required bool barrierDismissible,
     required UpgraderMessages messages,
   }) {
     showDialog(
