@@ -9,12 +9,7 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  // On Android, the default behavior will be to use the Google Play Store
-  // version of the app.
-  // On iOS, the default behavior will be to use the App Store version of
-  // the app, so update the Bundle Identifier in example/ios/Runner with a
-  // valid identifier already in the App Store.
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
@@ -25,9 +20,9 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Upgrader Example',
       home: Scaffold(
-        appBar: AppBar(title: Text('Upgrader Custom Card Example')),
+        appBar: AppBar(title: const Text('Upgrader Custom Card Example')),
         body: Container(
-          margin: EdgeInsets.only(left: 12.0, right: 12.0),
+          margin: const EdgeInsets.only(left: 12.0, right: 12.0),
           child: SingleChildScrollView(
             child: Column(
               children: [
@@ -44,7 +39,7 @@ class MyApp extends StatelessWidget {
     );
   }
 
-  Widget get _simpleCard => Card(
+  Widget get _simpleCard => const Card(
         child: SizedBox(
           width: 200,
           height: 50,
@@ -54,7 +49,7 @@ class MyApp extends StatelessWidget {
 }
 
 class MyUpgradeCard extends UpgradeCard {
-  MyUpgradeCard({super.upgrader});
+  MyUpgradeCard({super.key, super.upgrader});
 
   /// Override the [createState] method to provide a custom class
   /// with overridden methods.
@@ -80,7 +75,7 @@ class MyUpgradeCardState extends UpgradeCardState {
             },
           ),
         ],
-        content: Text(''),
+        content: const Text(''),
         title: Text(title ?? ''),
       ),
     );
