@@ -9,7 +9,7 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatefulWidget {
@@ -26,27 +26,27 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     final scaffold = Scaffold(
-      appBar: AppBar(title: Text('Upgrader Driver App')),
+      appBar: AppBar(title: const Text('Upgrader Driver App')),
       body: Center(
           child: Column(
         children: [
-          SizedBox(height: 32.0),
+          const SizedBox(height: 32.0),
           ElevatedButton(
             onPressed: () async {
               await Upgrader.clearSavedSettings();
               _upgrader = Upgrader(debugLogging: true);
               setState(() => _testState = 1);
             },
-            child: Text('Dialog Alert'),
+            child: const Text('Dialog Alert'),
           ),
-          SizedBox(height: 16.0),
+          const SizedBox(height: 16.0),
           ElevatedButton(
             onPressed: () async {
               await Upgrader.clearSavedSettings();
               _upgrader = Upgrader(debugLogging: true);
               setState(() => _testState = 2);
             },
-            child: Text('Dialog Alert - Cupertino'),
+            child: const Text('Dialog Alert - Cupertino'),
           ),
         ],
       )),
@@ -59,11 +59,11 @@ class _MyAppState extends State<MyApp> {
         break;
       case 1:
         content = UpgradeAlert(
-            key: Key('ua_1'), upgrader: _upgrader, child: scaffold);
+            key: const Key('ua_1'), upgrader: _upgrader, child: scaffold);
         break;
       case 2:
         content = UpgradeAlert(
-            key: Key('ua_2'),
+            key: const Key('ua_2'),
             upgrader: _upgrader,
             dialogStyle: UpgradeDialogStyle.cupertino,
             child: scaffold);
