@@ -10,28 +10,30 @@ void main() async {
   // Only call clearSavedSettings() during testing to reset internal values.
   await Upgrader.clearSavedSettings(); // REMOVE this for release builds
 
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      locale: Locale('ar'), // Arabic language shows right to left.
-      localizationsDelegates: [
+      locale: const Locale('ar'), // Arabic language shows right to left.
+      localizationsDelegates: const [
         GlobalMaterialLocalizations.delegate,
         GlobalWidgetsLocalizations.delegate,
         GlobalCupertinoLocalizations.delegate,
       ],
-      supportedLocales: [
-        const Locale('ar', ''), // Arabic, no country code
-        const Locale('he', ''), // Hebrew, no country code
+      supportedLocales: const [
+        Locale('ar', ''), // Arabic, no country code
+        Locale('he', ''), // Hebrew, no country code
       ],
       title: 'Upgrader Left to Right Example',
       home: UpgradeAlert(
           child: Scaffold(
-        appBar: AppBar(title: Text('Upgrader Left to Right Example')),
-        body: Center(child: Text('Checking...')),
+        appBar: AppBar(title: const Text('Upgrader Left to Right Example')),
+        body: const Center(child: Text('Checking...')),
       )),
     );
   }
