@@ -21,6 +21,9 @@ class ITunesSearchAPI {
   /// Provide an HTTP Client that can be replaced for mock testing.
   http.Client? client = http.Client();
 
+  /// Provide the HTTP headers used by [client].
+  Map<String, String>? clientHeaders;
+
   /// Enable print statements for debugging.
   bool debugLogging = false;
 
@@ -36,13 +39,14 @@ class ITunesSearchAPI {
     }
 
     final url = lookupURLByBundleId(bundleId,
-        country: country ??= '', useCacheBuster: useCacheBuster)!;
+        country: country ?? '', useCacheBuster: useCacheBuster)!;
     if (debugLogging) {
       print('upgrader: download: $url');
     }
 
     try {
-      final response = await client!.get(Uri.parse(url));
+      final response =
+          await client!.get(Uri.parse(url), headers: clientHeaders);
       if (debugLogging) {
         print('upgrader: response statusCode: ${response.statusCode}');
       }
@@ -73,7 +77,8 @@ class ITunesSearchAPI {
       print('upgrader: download: $url');
     }
     try {
-      final response = await client!.get(Uri.parse(url));
+      final response =
+          await client!.get(Uri.parse(url), headers: clientHeaders);
       final decodedResults = _decodeResults(response.body);
       return decodedResults;
     } catch (e) {
@@ -140,6 +145,7 @@ class ITunesSearchAPI {
             print(
                 'upgrader.ITunesSearchAPI: results are empty: $decodedResults');
           }
+          return null;
         }
         return decodedResults;
       }
