@@ -47,7 +47,9 @@ enum UpgraderMessage {
 ///   String get buttonTitleIgnore => 'My Ignore';
 /// }
 ///
-/// UpgradeAlert(messages: MyUpgraderMessages());
+/// final upgrader = Upgrader(messages: MyUpgraderMessages());
+/// ...
+/// UpgradeAlert(upgrader: upgrader);
 /// ```
 ///
 class UpgraderMessages {
@@ -89,13 +91,11 @@ class UpgraderMessages {
     Locale? locale;
     if (context != null) {
       locale = Localizations.maybeLocaleOf(context);
-    } else {
-      // Get the system locale
-      locale = PlatformDispatcher.instance.locale;
     }
-    final code = locale == null || locale.languageCode.isEmpty
-        ? 'en'
-        : locale.languageCode;
+    // Get the system locale
+    locale ??= PlatformDispatcher.instance.locale;
+
+    final code = locale.languageCode.isEmpty ? 'en' : locale.languageCode;
     return code;
   }
 
@@ -161,7 +161,7 @@ class UpgraderMessages {
         break;
       case 'ht':
         message =
-            'Yon nouvo vèsyon {{appName}} disponib! Vèsyon {{currentAppStoreVersion}} disponib, epi ou gen vèsyon {{currentInstalledVersion}}.';
+            'Yon nouvo vèsyon {{appName}} disponib! Vèsyon {{currentAppStoreVersion}} disponib, ou gen vèsyon {{currentInstalledVersion}}.';
         break;
       case 'hu':
         message =
@@ -190,6 +190,10 @@ class UpgraderMessages {
       case 'ko':
         message =
             '{{appName}}이 새 버전으로 업데이트되었습니다! 최신 버전 {{currentAppStoreVersion}}으로 업그레이드 가능합니다 - 현재 버전 {{currentInstalledVersion}}.';
+        break;
+      case 'ku':
+        message =
+            'وەشانی نوێی {{appName}} بەردەستە! وەشانی {{currentAppStoreVersion}} بەردەستە- تۆ وەشانی {{currentInstalledVersion}} دابەزاندوە.';
         break;
       case 'lt':
         message =
@@ -297,7 +301,7 @@ class UpgraderMessages {
         message = 'नज़रअंदाज़ करना';
         break;
       case 'ht':
-        message = 'IGNORE';
+        message = 'INYORE';
         break;
       case 'hu':
         message = 'KIHAGYOM';
@@ -319,6 +323,9 @@ class UpgraderMessages {
         break;
       case 'ko':
         message = '무시';
+        break;
+      case 'ku':
+        message = 'پشتگوێخستن';
         break;
       case 'lt':
         message = 'IGNORUOTI';
@@ -439,6 +446,9 @@ class UpgraderMessages {
       case 'ko':
         message = '나중에';
         break;
+      case 'ku':
+        message = 'دواتر';
+        break;
       case 'lt':
         message = 'ATNAUJINTI VĖLIAU';
         break;
@@ -555,6 +565,9 @@ class UpgraderMessages {
       case 'ko':
         message = '지금 업데이트';
         break;
+      case 'ku':
+        message = 'نوێکردنەوە';
+        break;
       case 'lt':
         message = 'ATNAUJINTI DABAR';
         break;
@@ -648,7 +661,7 @@ class UpgraderMessages {
         message = 'क्या आप इसे अभी अद्यतन करना चाहेंगे?';
         break;
       case 'ht':
-        message = 'Èske ou vle mete ajou aplikasyon an kounye a?';
+        message = 'Èske ou vle mete aplikasyon an ajou kounye a?';
         break;
       case 'hu':
         message = 'Akarja most frissíteni?';
@@ -670,6 +683,9 @@ class UpgraderMessages {
         break;
       case 'ko':
         message = '지금 업데이트를 시작하시겠습니까?';
+        break;
+      case 'ku':
+        message = 'دەتەوێت ئێستا نوێی بکەیەوە؟';
         break;
       case 'lt':
         message = 'Ar norite atnaujinti dabar?';
@@ -760,6 +776,9 @@ class UpgraderMessages {
       case 'ja':
         message = 'リリースノート';
         break;
+      case 'ku':
+        message = 'تیبینەکانی وەشان';
+        break;
       case 'pt':
         message = 'Novidades';
         break;
@@ -845,7 +864,7 @@ class UpgraderMessages {
         message = 'अद्यतन ऐप?';
         break;
       case 'ht':
-        message = 'Mete ajou app a?';
+        message = 'Mete app la ajou?';
         break;
       case 'hu':
         message = 'FrissÍtés?';
@@ -867,6 +886,9 @@ class UpgraderMessages {
         break;
       case 'ko':
         message = '앱을 업데이트하시겠습니까?';
+        break;
+      case 'ku':
+        message = 'نوێکردنەوەی ئەپ؟';
         break;
       case 'lt':
         message = 'Atnaujinti programą?';
